amex_claim_dashboard/requirements.txt
streamlit==1.38.0
pandas==2.2.2
python-dateutil==2.9.0.post0
Jinja2==3.1.4
reportlab==4.2.2
beautifulsoup4==4.12.3
lxml==5.3.0
